[Selection] move selection controller to subfolder
